requirements.txt
matplotlib==3.5.1
numpy==1.21.5
pandas==1.2.4
Pillow==9.0.1
Pillow==10.1.0
scikit_learn==1.0.2
seaborn==0.13.0
torch==1.12.1
torchvision==0.13.1
